refactor(discord): split callbacks.ts into callback modules
callbacks.ts was 684 lines covering component interactions, modal
submits, and autocomplete with a module-level autocomplete cache. Split
into a callbacks/ folder mirroring the telegram callbacks/ layout:
  - shared       — ComponentInteraction type + chatIdFromInteraction
  - components   — button/select handler + refreshSettingsPanel + forwardToAgent
  - modals       — modal submissions (pulse interval)
  - autocomplete — /model name autocomplete + cached model list

index.ts re-exports handleComponentInteraction / handleModalSubmit /
handleAutocomplete. The router's dynamic imports updated to the barrel.
